requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.9.1
bios==0.1.2
cachier==1.5.0
certifi==2019.9.11
chardet==3.0.4
chromedriver-binary==72.0.3626.69.0
configparser==4.0.2
contextlib2==0.6.0.post1
coverage==4.5.4
dropbox==9.4.0
et-xmlfile==1.0.1
funcsigs==1.0.2
html5lib==1.1
idna==2.8
importlib-metadata==0.23
iniconfig==1.1.1
jdcal==1.4.1
lxml
#lxml==4.6.3
more-itertools==5.0.0
#numpy==1.16.5
openpyxl==3.0.5
oyaml==1.0
packaging==19.2
pandas
#pandas==0.24.2
pathlib2==2.3.5
pathtools==0.1.2
pluggy==0.13.0
portalocker==2.0.0
pretty-html-table==0.9.dev0
py==1.10.0
py-w3c==0.3.1
PyJWT==1.7.1
pyparsing==2.4.2
pytest==6.2.2
pytest-cov==2.8.1
pytest-mock==3.5.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.4
requests==2.25.1
requests-futures==1.0.0
scandir==1.10.0
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.5
tabulate
#tabulate==0.8.6
toml==0.10.2
tqdm==4.54.1
urllib3==1.26.5
watchdog==1.0.2
wcwidth==0.1.7
webencodings==0.5.1
yahooquery==2.2.14
zipp==0.6.0
tabula-py
pypdf2
